+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0-SNAPSHOT - 06/30/16
 
 ### Added
+- Signup and login JUnit tests
+- Parse signup and login function
 - Stub method of Login
 - Initialize with App Id and Rest API key